% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jmjax.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ranef}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nlme}{\code{\link[nlme:ranef]{ranef()}}}
}}

