% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jmjax.R
\name{print.jmjax}
\alias{print.jmjax}
\title{Print a jmjax fit}
\usage{
\method{print}{jmjax}(x, ...)
}
\arguments{
\item{x}{A \code{"jmjax"} fit object.}

\item{...}{Further arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a jmjax fit
}
